// src/auth/interface/roles.enum.ts
export enum Role {
  NationalAdmin = 'National Admin',
  NationalSCM = 'National Scm',
  RegionalAdmin = 'Regional Admin',
  RegionalSCM = 'Regional Scm',
  HospitalAdmin = 'Hospital Admin',
  HospitalSCM = 'Hospital Scm',
  DeptHead = 'Dept Head',
  HealthcareWorker = 'Healthcare Worker',
  SupplierVendor = 'Supplier Vendor',
}